# Conflicts:
#	backend/src/main/java/com/apink/poppin/api/popup/controller/PopupController.java

+++ backend/src/main/java/com/apink/poppin/api/reservation/entity/ReservationStatement.java
@@ -16,7 +16,7 @@ public class ReservationStatement {
     @Id
     @NotNull
     @Column(nullable = false)
-    private Integer reservationStatementId;
+    private Long reservationStatementId;
 
     @NotNull
     @Size(max = 16)
